refactor(api-v2): move invite logic to shared service and use in controller
